Add spacing rules for directive block comments and refactor related methods
