+++ src/main/kotlin/org/domaframework/doma/intellij/formatter/block/SqlBlock.kt
@@ -247,6 +247,161 @@ open class SqlBlock(
     ): Spacing? = null
 
     /**
+     * Creates a spacing builder specifically for directive block comments.
+     */
+    protected fun createBlockDirectiveCommentSpacingBuilder(): SqlCustomSpacingBuilder =
+        SqlCustomSpacingBuilder()
+            .withSpacing(
+                SqlTypes.BLOCK_COMMENT_START,
+                SqlTypes.EL_ID_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BLOCK_COMMENT_START,
+                SqlTypes.EL_PRIMARY_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BLOCK_COMMENT_START,
+                SqlTypes.EL_STRING,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BLOCK_COMMENT_START,
+                SqlTypes.EL_NUMBER,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BLOCK_COMMENT_START,
+                SqlTypes.BOOLEAN,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BLOCK_COMMENT_START,
+                SqlTypes.EL_NULL,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BLOCK_COMMENT_START,
+                SqlTypes.EL_FIELD_ACCESS_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BLOCK_COMMENT_START,
+                SqlTypes.EL_STATIC_FIELD_ACCESS_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BLOCK_COMMENT_START,
+                SqlTypes.HASH,
+                Spacing.createSpacing(0, 0, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.HASH,
+                SqlTypes.EL_ID_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.HASH,
+                SqlTypes.EL_PRIMARY_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.HASH,
+                SqlTypes.EL_STRING,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.HASH,
+                SqlTypes.EL_NUMBER,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.HASH,
+                SqlTypes.BOOLEAN,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.HASH,
+                SqlTypes.EL_NULL,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.HASH,
+                SqlTypes.EL_FIELD_ACCESS_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.HASH,
+                SqlTypes.EL_STATIC_FIELD_ACCESS_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BLOCK_COMMENT_START,
+                SqlTypes.CARET,
+                Spacing.createSpacing(0, 0, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.CARET,
+                SqlTypes.EL_ID_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.CARET,
+                SqlTypes.EL_PRIMARY_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.CARET,
+                SqlTypes.EL_STRING,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.CARET,
+                SqlTypes.EL_NUMBER,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.CARET,
+                SqlTypes.BOOLEAN,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.CARET,
+                SqlTypes.EL_NULL,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.CARET,
+                SqlTypes.EL_FIELD_ACCESS_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.CARET,
+                SqlTypes.EL_STATIC_FIELD_ACCESS_EXPR,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BLOCK_COMMENT_CONTENT,
+                SqlTypes.BLOCK_COMMENT_END,
+                Spacing.createSpacing(0, 0, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.EL_FIELD_ACCESS_EXPR,
+                SqlTypes.OTHER,
+                Spacing.createSpacing(1, 1, 0, false, 0),
+            ).withSpacing(
+                SqlTypes.EL_STATIC_FIELD_ACCESS_EXPR,
+                SqlTypes.OTHER,
+                Spacing.createSpacing(1, 1, 0, false, 0),
+            ).withSpacing(
+                SqlTypes.EL_ID_EXPR,
+                SqlTypes.BLOCK_COMMENT_END,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.EL_PRIMARY_EXPR,
+                SqlTypes.BLOCK_COMMENT_END,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.STRING,
+                SqlTypes.BLOCK_COMMENT_END,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.EL_NUMBER,
+                SqlTypes.BLOCK_COMMENT_END,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.EL_NULL,
+                SqlTypes.BLOCK_COMMENT_END,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.BOOLEAN,
+                SqlTypes.BLOCK_COMMENT_END,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.EL_FIELD_ACCESS_EXPR,
+                SqlTypes.BLOCK_COMMENT_END,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            ).withSpacing(
+                SqlTypes.EL_STATIC_FIELD_ACCESS_EXPR,
+                SqlTypes.BLOCK_COMMENT_END,
+                Spacing.createSpacing(1, 1, 0, true, 0),
+            )
+
+    /**
      * Returns the child indentation for the block.
      *
      * @return The indent to apply to child blocks
