+++ src/main/kotlin/org/domaframework/doma/intellij/formatter/block/comment/SqlElBlockCommentBlock.kt
@@ -102,6 +102,16 @@ open class SqlElBlockCommentBlock(
             ->
                 SqlOperationBlock(child, context)
 
+            SqlTypes.EL_EQ_EXPR, SqlTypes.EL_NE_EXPR, SqlTypes.EL_GE_EXPR, SqlTypes.EL_GT_EXPR, SqlTypes.EL_LE_EXPR, SqlTypes.EL_LT_EXPR,
+            SqlTypes.EL_AND_EXPR, SqlTypes.EL_OR_EXPR, SqlTypes.EL_NOT_EXPR,
+            SqlTypes.EL_ADD_EXPR, SqlTypes.EL_SUBTRACT_EXPR, SqlTypes.EL_MULTIPLY_EXPR, SqlTypes.EL_DIVIDE_EXPR, SqlTypes.EL_MOD_EXPR,
+            ->
+                SqlElBlockCommentBlock(
+                    child,
+                    context,
+                    createBlockDirectiveCommentSpacingBuilder(),
+                )
+
             SqlTypes.EL_FIELD_ACCESS_EXPR ->
                 SqlElFieldAccessBlock(
                     child,
