+++ src/main/kotlin/org/domaframework/doma/intellij/formatter/builder/SqlFormattingModelBuilder.kt
@@ -102,7 +102,31 @@ class SqlFormattingModelBuilder : FormattingModelBuilder {
             .spacing(1, 1, 0, false, 0)
             .around(SqlTypes.SLASH)
             .spacing(1, 1, 0, false, 0)
+            .around(SqlTypes.PERCENT)
+            .spacing(1, 1, 0, false, 0)
             .before(SqlTypes.FUNCTION_NAME)
+            .spacing(1, 1, 0, false, 0)
+            .around(SqlTypes.SEPARATOR)
+            .spacing(1, 1, 0, false, 0)
+            .before(SqlTypes.EL_NOT)
+            .spacing(1, 1, 0, false, 0)
+            .after(SqlTypes.EL_NOT)
+            .spacing(0, 0, 0, false, 0)
+            .around(SqlTypes.EL_AND)
+            .spacing(1, 1, 0, false, 0)
+            .around(SqlTypes.EL_OR)
+            .spacing(1, 1, 0, false, 0)
+            .around(SqlTypes.EL_EQ)
+            .spacing(1, 1, 0, false, 0)
+            .around(SqlTypes.GE)
+            .spacing(1, 1, 0, false, 0)
+            .around(SqlTypes.GT)
+            .spacing(1, 1, 0, false, 0)
+            .around(SqlTypes.LE)
+            .spacing(1, 1, 0, false, 0)
+            .around(SqlTypes.LT)
+            .spacing(1, 1, 0, false, 0)
+            .around(SqlTypes.EL_NE)
             .spacing(1, 1, 0, false, 0)
 
     private fun createCustomSpacingBuilder(): SqlCustomSpacingBuilder =
